♻️refactor: change justification type faker path
